Intense rework for the new tag structure. There is now a tag graph which is distinct from the delta graph. It probably should be a lod, but I did it this way.  We'll see how long it lasts.
bk: 3a288cf4LOYaxcqNfBd1aLE5dmL-yw

+++ src/resolve.c
@@ -2692,7 +2692,7 @@ resolve_cleanup(opts *opts, int what)
 	}
 	
 	if (what & CLEAN_PENDING) {
-		if (pendingFile[0]) {
+		if (pendingFile[0] && !getenv("TAKEPATCH_SAVEDIRS")) {
 			unlink(pendingFile);
 		}
 		rmdir(ROOT2PENDING);
